Added interaction matrix to mult-mono kernel
This is supposed to reduce the number of redundant computations in this interaction method.
Before, the interaction was calculated for all cells within the stencil, not only for the
neighbors. This did not change the results since these interaction calls were evaluated to
zero anyway, but it reduced the performance. This commit will not solve the issue completely,
it is meant as a temporary bandaid.