Improved cache usage in p2p interaction kernel
The previous blocking that was used in the monopole-monopole interaction kernel was copied
from the multipole-multipole interaction kernel and thus not really suited for the
monopole-monopole interactions since they require way less data per iteration of the inner
loop. Thus we can increase the blocking size to better utilize the cache.