Add option for setting the cuda arch for clang builds
Not happy with this, but it works for now!
Hopefully CMake adds a way to auto detect the platform.

For kokkos builds we could just rely on their CUDA arch, but it is a bit
of a pain as we redefine compiler flags for a few files and thus would
need to convert the Kokkos CMake Variables into the actual compiler
flags ourselves.