Replace custom print function with printf
The custom function caused a ton of format-security warning.
Furthermore, it is not really necessary as its only job is to flush
after the printing, which can also be done using setbuf once at the
beginning of the program.

Basically, the custom print is causing warnings, is not necessary, and
complicates the code. Hence, this commit removes it.