Quick workaround to re-enable GPU compilation with smaller thetas
Renders the monopole CUDA kernel inoperational, however, KOKKOS_CUDA
still works. CUDA will throw an exception when someone tries to use it
in this configuration!

For the future, it would be handy to have a CMAKE variable to disable
CUDA at compiletime but not KOKKOS_CUDA. This workaround should serve
for now.